[housekeeping] Restructure pythonic resource tests
