+++ python_modules/dagster/dagster_tests/core_tests/resource_tests/pythonic_resources/test_general_pythonic_resources.py
@@ -1,59 +1,40 @@
 import enum
 import json
-import os
-import re
-import subprocess
 import sys
-import tempfile
 from abc import ABC, abstractmethod
-from typing import Any, Callable, List, Mapping, Optional, cast
+from typing import Any, Callable, List, Mapping, Optional
 
 import mock
 import pytest
 from dagster import (
-    BindResourcesToJobs,
-    DagsterInstance,
-    IOManager,
-    JobDefinition,
-    RunRequest,
-    ScheduleDefinition,
-    asset,
-    job,
-    op,
-    repository,
-    resource,
-    sensor,
-)
-from dagster._check import CheckError
-from dagster._config.field import Field
-from dagster._config.field_utils import EnvVar
-from dagster._config.pythonic_config import (
     Config,
-    ConfigurableIOManager,
     ConfigurableIOManagerFactory,
     ConfigurableLegacyIOManagerAdapter,
     ConfigurableResource,
-    ConfigurableResourceFactory,
+    DagsterInstance,
+    Definitions,
+    Field,
     IAttachDifferentObjectToOpContext,
+    InitResourceContext,
+    IOManager,
+    IOManagerDefinition,
+    OpExecutionContext,
     ResourceDependency,
+    ResourceParam,
+    RunConfig,
+    asset,
+    build_init_resource_context,
+    io_manager,
+    job,
+    op,
+    resource,
 )
+from dagster._check import CheckError
+from dagster._config.pythonic_config import ConfigurableResourceFactory
 from dagster._core.definitions.assets_job import build_assets_job
-from dagster._core.definitions.definitions_class import Definitions
-from dagster._core.definitions.repository_definition.repository_data_builder import (
-    build_caching_repository_data_from_dict,
-)
-from dagster._core.definitions.resource_annotation import ResourceParam
-from dagster._core.definitions.run_config import RunConfig
 from dagster._core.errors import (
-    DagsterInvalidConfigError,
     DagsterInvalidDefinitionError,
-    DagsterInvalidInvocationError,
 )
-from dagster._core.execution.context.compute import OpExecutionContext
-from dagster._core.execution.context.init import InitResourceContext, build_init_resource_context
-from dagster._core.execution.context.invocation import build_op_context
-from dagster._core.storage.io_manager import IOManagerDefinition, io_manager
-from dagster._core.test_utils import environ
 from dagster._utils.cached_method import cached_method
 from pydantic import (
     Field as PyField,
@@ -909,150 +890,6 @@ def test_nested_function_resource_runtime_config():
     assert out_txt == ["msg: foo!", "msg: bar!"]
 
 
-def get_pyright_reveal_type_output(filename) -> List[str]:
-    stdout = subprocess.check_output(["pyright", filename]).decode("utf-8")
-    match = re.findall(r'Type of "(?:[^"]+)" is "([^"]+)"', stdout)
-    assert match
-    return match
-
-
-def get_mypy_type_output(filename) -> List[str]:
-    stdout = subprocess.check_output(["mypy", filename]).decode("utf-8")
-    match = re.findall(r'note: Revealed type is "([^"]+)"', stdout)
-    assert match
-    return match
-
-
-# you can specify the -m "typesignature" flag to run these tests, they're
-# slow so we don't want to run them by default
-@pytest.mark.typesignature
-def test_type_signatures_constructor_nested_resource():
-    with tempfile.TemporaryDirectory() as tempdir:
-        filename = os.path.join(tempdir, "test.py")
-
-        with open(filename, "w") as f:
-            f.write(
-                """
-from dagster import ConfigurableResource
-
-class InnerResource(ConfigurableResource):
-    a_string: str
-
-class OuterResource(ConfigurableResource):
-    inner: InnerResource
-    a_bool: bool
-
-reveal_type(InnerResource.__init__)
-reveal_type(OuterResource.__init__)
-
-my_outer = OuterResource(inner=InnerResource(a_string="foo"), a_bool=True)
-reveal_type(my_outer.inner)
-"""
-            )
-
-        pyright_out = get_pyright_reveal_type_output(filename)
-        mypy_out = get_mypy_type_output(filename)
-
-        # Ensure constructor signature is correct (mypy doesn't yet support Pydantic model constructor type hints)
-        assert pyright_out[0] == "(self: InnerResource, *, a_string: str) -> None"
-        assert (
-            pyright_out[1]
-            == "(self: OuterResource, *, inner: InnerResource | PartialResource[InnerResource],"
-            " a_bool: bool) -> None"
-        )
-
-        # Ensure that the retrieved type is the same as the type of the resource (no partial)
-        assert pyright_out[2] == "InnerResource"
-        assert mypy_out[2] == "test.InnerResource"
-
-
-@pytest.mark.typesignature
-def test_type_signatures_config_at_launch():
-    with tempfile.TemporaryDirectory() as tempdir:
-        filename = os.path.join(tempdir, "test.py")
-
-        with open(filename, "w") as f:
-            f.write(
-                """
-from dagster import ConfigurableResource
-
-class MyResource(ConfigurableResource):
-    a_string: str
-
-reveal_type(MyResource.configure_at_launch())
-"""
-            )
-
-        pyright_out = get_pyright_reveal_type_output(filename)
-        mypy_out = get_mypy_type_output(filename)
-
-        # Ensure partial resource is correctly parameterized
-        assert pyright_out[0] == "PartialResource[MyResource]"
-        assert mypy_out[0].endswith("PartialResource[test.MyResource]")
-
-
-@pytest.mark.typesignature
-def test_type_signatures_constructor_resource_dependency():
-    with tempfile.TemporaryDirectory() as tempdir:
-        filename = os.path.join(tempdir, "test.py")
-
-        with open(filename, "w") as f:
-            f.write(
-                """
-from dagster import ConfigurableResource, ResourceDependency
-
-class StringDependentResource(ConfigurableResource):
-    a_string: ResourceDependency[str]
-
-reveal_type(StringDependentResource.__init__)
-
-my_str_resource = StringDependentResource(a_string="foo")
-reveal_type(my_str_resource.a_string)
-"""
-            )
-
-        pyright_out = get_pyright_reveal_type_output(filename)
-        mypy_out = get_mypy_type_output(filename)
-
-        # Ensure constructor signature supports str Resource, PartialResource, raw str, or a
-        # resource function that returns a str
-        assert (
-            pyright_out[0]
-            == "(self: StringDependentResource, *, a_string: ConfigurableResourceFactory[str] |"
-            " PartialResource[str] | ResourceDefinition | str) -> None"
-        )
-
-        # Ensure that the retrieved type is str
-        assert pyright_out[1] == "str"
-        assert mypy_out[1] == "builtins.str"
-
-
-@pytest.mark.typesignature
-def test_type_signatures_alias():
-    with tempfile.TemporaryDirectory() as tempdir:
-        filename = os.path.join(tempdir, "test.py")
-
-        with open(filename, "w") as f:
-            f.write(
-                """
-from dagster import ConfigurableResource
-from pydantic import Field
-
-class ResourceWithAlias(ConfigurableResource):
-    _schema: str = Field(alias="schema")
-
-reveal_type(ResourceWithAlias.__init__)
-
-my_resource = ResourceWithAlias(schema="foo")
-"""
-            )
-
-        pyright_out = get_pyright_reveal_type_output(filename)
-
-        # Ensure constructor signature shows schema as the alias
-        assert pyright_out[0] == "(self: ResourceWithAlias, *, schema: str) -> None"
-
-
 def test_nested_config_class() -> None:
     # Validate that we can nest Config classes in a pythonic resource
 
@@ -1089,222 +926,6 @@ def test_nested_config_class() -> None:
 
     assert defs.get_implicit_global_asset_job_def().execute_in_process().success
     assert executed["yes"]
-
-
-def test_env_var():
-    with environ(
-        {
-            "ENV_VARIABLE_FOR_TEST": "SOME_VALUE",
-        }
-    ):
-
-        class ResourceWithString(ConfigurableResource):
-            a_str: str
-
-        executed = {}
-
-        @asset
-        def an_asset(a_resource: ResourceWithString):
-            assert a_resource.a_str == "SOME_VALUE"
-            executed["yes"] = True
-
-        defs = Definitions(
-            assets=[an_asset],
-            resources={
-                "a_resource": ResourceWithString(
-                    a_str=EnvVar("ENV_VARIABLE_FOR_TEST"),
-                )
-            },
-        )
-
-        assert defs.get_implicit_global_asset_job_def().execute_in_process().success
-        assert executed["yes"]
-
-
-def test_env_var_data_structure() -> None:
-    with environ(
-        {
-            "FOO": "hello",
-            "BAR": "world",
-        }
-    ):
-
-        class ResourceWithString(ConfigurableResource):
-            my_list: List[str]
-            my_dict: Mapping[str, str]
-
-        executed = {}
-
-        @asset
-        def an_asset(a_resource: ResourceWithString):
-            assert len(a_resource.my_list) == 2
-            assert a_resource.my_list[0] == "hello"
-            assert a_resource.my_list[1] == "world"
-            assert len(a_resource.my_dict) == 2
-            assert a_resource.my_dict["foo"] == "hello"
-            assert a_resource.my_dict["bar"] == "world"
-            executed["yes"] = True
-
-        defs = Definitions(
-            assets=[an_asset],
-            resources={
-                "a_resource": ResourceWithString(
-                    my_list=[EnvVar("FOO"), EnvVar("BAR")],
-                    my_dict={
-                        "foo": EnvVar("FOO"),
-                        "bar": EnvVar("BAR"),
-                    },
-                )
-            },
-        )
-
-        assert defs.get_implicit_global_asset_job_def().execute_in_process().success
-        assert executed["yes"]
-
-
-def test_runtime_config_env_var():
-    out_txt = []
-
-    class WriterResource(ConfigurableResource):
-        prefix: str
-
-        def output(self, text: str) -> None:
-            out_txt.append(f"{self.prefix}{text}")
-
-    @asset
-    def hello_world_asset(writer: WriterResource):
-        writer.output("hello, world!")
-
-    defs = Definitions(
-        assets=[hello_world_asset],
-        resources={"writer": WriterResource.configure_at_launch()},
-    )
-
-    os.environ["MY_PREFIX_FOR_TEST"] = "greeting: "
-    try:
-        assert (
-            defs.get_implicit_global_asset_job_def()
-            .execute_in_process(
-                {"resources": {"writer": {"config": {"prefix": EnvVar("MY_PREFIX_FOR_TEST")}}}}
-            )
-            .success
-        )
-        assert out_txt == ["greeting: hello, world!"]
-    finally:
-        del os.environ["MY_PREFIX_FOR_TEST"]
-
-
-def test_env_var_err():
-    if "UNSET_ENV_VAR" in os.environ:
-        del os.environ["UNSET_ENV_VAR"]
-
-    class ResourceWithString(ConfigurableResource):
-        a_str: str
-
-    @asset
-    def an_asset(a_resource: ResourceWithString):
-        pass
-
-    # No error constructing the resource, only at runtime
-    defs = Definitions(
-        assets=[an_asset],
-        resources={
-            "a_resource": ResourceWithString(
-                a_str=EnvVar("UNSET_ENV_VAR"),
-            )
-        },
-    )
-    with pytest.raises(
-        DagsterInvalidConfigError,
-        match=(
-            'You have attempted to fetch the environment variable "UNSET_ENV_VAR" which is not set.'
-        ),
-    ):
-        defs.get_implicit_global_asset_job_def().execute_in_process()
-
-    # Test using runtime configuration of the resource
-    defs = Definitions(
-        assets=[an_asset],
-        resources={"a_resource": ResourceWithString.configure_at_launch()},
-    )
-    with pytest.raises(
-        DagsterInvalidConfigError,
-        match=(
-            'You have attempted to fetch the environment variable "UNSET_ENV_VAR_RUNTIME" which is'
-            " not set."
-        ),
-    ):
-        defs.get_implicit_global_asset_job_def().execute_in_process(
-            {"resources": {"a_resource": {"config": {"a_str": EnvVar("UNSET_ENV_VAR_RUNTIME")}}}}
-        )
-
-
-def test_env_var_nested_resources() -> None:
-    class ResourceWithString(ConfigurableResource):
-        a_str: str
-
-    class OuterResource(ConfigurableResource):
-        inner: ResourceWithString
-
-    executed = {}
-
-    @asset
-    def an_asset(a_resource: OuterResource):
-        assert a_resource.inner.a_str == "SOME_VALUE"
-        executed["yes"] = True
-
-    defs = Definitions(
-        assets=[an_asset],
-        resources={
-            "a_resource": OuterResource(
-                inner=ResourceWithString(
-                    a_str=EnvVar("ENV_VARIABLE_FOR_TEST"),
-                )
-            )
-        },
-    )
-
-    with environ(
-        {
-            "ENV_VARIABLE_FOR_TEST": "SOME_VALUE",
-        }
-    ):
-        assert defs.get_implicit_global_asset_job_def().execute_in_process().success
-        assert executed["yes"]
-
-
-def test_env_var_nested_config() -> None:
-    class NestedWithString(Config):
-        a_str: str
-
-    class OuterResource(ConfigurableResource):
-        inner: NestedWithString
-
-    executed = {}
-
-    @asset
-    def an_asset(a_resource: OuterResource):
-        assert a_resource.inner.a_str == "SOME_VALUE"
-        executed["yes"] = True
-
-    defs = Definitions(
-        assets=[an_asset],
-        resources={
-            "a_resource": OuterResource(
-                inner=NestedWithString(
-                    a_str=EnvVar("ENV_VARIABLE_FOR_TEST"),
-                )
-            )
-        },
-    )
-
-    with environ(
-        {
-            "ENV_VARIABLE_FOR_TEST": "SOME_VALUE",
-        }
-    ):
-        assert defs.get_implicit_global_asset_job_def().execute_in_process().success
-        assert executed["yes"]
 
 
 def test_using_enum() -> None:
@@ -1484,180 +1105,6 @@ def test_extending_resource_nesting() -> None:
     assert executed["yes"]
 
 
-def test_bind_resource_to_job_at_defn_time_err() -> None:
-    out_txt = []
-
-    class WriterResource(ConfigurableResource):
-        prefix: str
-
-        def output(self, text: str) -> None:
-            out_txt.append(f"{self.prefix}{text}")
-
-    @op
-    def hello_world_op(writer: WriterResource):
-        writer.output("hello, world!")
-
-    @job
-    def hello_world_job():
-        hello_world_op()
-
-    # Validate that jobs without bound resources error at repository construction time
-    with pytest.raises(
-        DagsterInvalidDefinitionError,
-        match="resource with key 'writer' required by op 'hello_world_op' was not provided",
-    ):
-        build_caching_repository_data_from_dict({"jobs": {"hello_world_job": hello_world_job}})
-
-    with pytest.raises(
-        DagsterInvalidDefinitionError,
-        match="resource with key 'writer' required by op 'hello_world_op' was not provided",
-    ):
-
-        @repository
-        def my_repo():
-            return [hello_world_job]
-
-    # Validate that this also happens with Definitions
-    with pytest.raises(
-        DagsterInvalidDefinitionError,
-        match="resource with key 'writer' required by op 'hello_world_op' was not provided",
-    ):
-        Definitions(
-            jobs=[hello_world_job],
-        )
-
-
-def test_bind_resource_to_job_at_defn_time() -> None:
-    out_txt = []
-
-    class WriterResource(ConfigurableResource):
-        prefix: str
-
-        def output(self, text: str) -> None:
-            out_txt.append(f"{self.prefix}{text}")
-
-    @op
-    def hello_world_op(writer: WriterResource):
-        writer.output("hello, world!")
-
-    @job
-    def hello_world_job():
-        hello_world_op()
-
-    # Bind the resource to the job at definition time and validate that it works
-    defs = Definitions(
-        jobs=[hello_world_job],
-        resources={
-            "writer": WriterResource(prefix=""),
-        },
-    )
-
-    assert defs.get_job_def("hello_world_job").execute_in_process().success
-    assert out_txt == ["hello, world!"]
-
-    out_txt.clear()
-
-    defs = Definitions(
-        jobs=[hello_world_job],
-        resources={
-            "writer": WriterResource(prefix="msg: "),
-        },
-    )
-
-    assert defs.get_job_def("hello_world_job").execute_in_process().success
-    assert out_txt == ["msg: hello, world!"]
-
-
-def test_bind_resource_to_job_at_defn_time_bind_resources_to_jobs() -> None:
-    out_txt = []
-
-    class WriterResource(ConfigurableResource):
-        prefix: str
-
-        def output(self, text: str) -> None:
-            out_txt.append(f"{self.prefix}{text}")
-
-    @op
-    def hello_world_op(writer: WriterResource):
-        writer.output("hello, world!")
-
-    @job
-    def hello_world_job():
-        hello_world_op()
-
-    # BindResourcesToJobs is a no-op now
-    defs = Definitions(
-        jobs=BindResourcesToJobs([hello_world_job]),
-        resources={
-            "writer": WriterResource(prefix=""),
-        },
-    )
-
-    assert defs.get_job_def("hello_world_job").execute_in_process().success
-    assert out_txt == ["hello, world!"]
-
-    out_txt.clear()
-
-    # BindResourcesToJobs is a no-op now
-    defs = Definitions(
-        jobs=BindResourcesToJobs([hello_world_job]),
-        resources={
-            "writer": WriterResource(prefix="msg: "),
-        },
-    )
-
-    assert defs.get_job_def("hello_world_job").execute_in_process().success
-    assert out_txt == ["msg: hello, world!"]
-
-
-def test_bind_resource_to_job_with_job_config() -> None:
-    out_txt = []
-
-    class WriterResource(ConfigurableResource):
-        prefix: str
-
-        def output(self, text: str) -> None:
-            out_txt.append(f"{self.prefix}{text}")
-
-    class OpConfig(Config):
-        message: str = "hello, world!"
-
-    @op
-    def hello_world_op(writer: WriterResource, config: OpConfig):
-        writer.output(config.message)
-
-    @job(config={})
-    def hello_world_job() -> None:
-        hello_world_op()
-
-    @job(config={"ops": {"hello_world_op": {"config": {"message": "hello, earth!"}}}})
-    def hello_earth_job() -> None:
-        hello_world_op()
-
-    defs = Definitions(
-        jobs=[hello_world_job, hello_earth_job],
-        resources={
-            "writer": WriterResource(prefix="msg: "),
-        },
-    )
-
-    assert defs.get_job_def("hello_world_job").execute_in_process().success
-    assert out_txt == ["msg: hello, world!"]
-    out_txt.clear()
-
-    assert defs.get_job_def("hello_earth_job").execute_in_process().success
-    assert out_txt == ["msg: hello, earth!"]
-
-    # Validate that we correctly error
-    with pytest.raises(
-        DagsterInvalidDefinitionError,
-        match="resource with key 'writer' required by op 'hello_world_op' was not provided",
-    ):
-        Definitions(
-            jobs=[hello_world_job],
-        )
-
-
 def test_execute_in_process() -> None:
     out_txt = []
 
@@ -1688,223 +1135,6 @@ def test_execute_in_process() -> None:
         resources={"writer": WriterResource(prefix="msg: ")}
     ).success
     assert out_txt == ["msg: hello, world!"]
-
-
-def test_bind_resource_to_job_at_defn_time_override() -> None:
-    out_txt = []
-
-    class WriterResource(ConfigurableResource):
-        prefix: str
-
-        def output(self, text: str) -> None:
-            out_txt.append(f"{self.prefix}{text}")
-
-    @op
-    def hello_world_op(writer: WriterResource):
-        writer.output("hello, world!")
-
-    # Binding the resource to the job at definition time should not override the resource
-    @job(
-        resource_defs={
-            "writer": WriterResource(prefix="job says: "),
-        }
-    )
-    def hello_world_job_with_override():
-        hello_world_op()
-
-    @job
-    def hello_world_job_no_override():
-        hello_world_op()
-
-    defs = Definitions(
-        jobs=[hello_world_job_with_override, hello_world_job_no_override],
-        resources={
-            "writer": WriterResource(prefix="definitions says: "),
-        },
-    )
-
-    assert defs.get_job_def("hello_world_job_with_override").execute_in_process().success
-    assert out_txt == ["job says: hello, world!"]
-    out_txt.clear()
-
-    assert defs.get_job_def("hello_world_job_no_override").execute_in_process().success
-    assert out_txt == ["definitions says: hello, world!"]
-
-
-def test_bind_resource_to_instigator() -> None:
-    out_txt = []
-
-    class WriterResource(ConfigurableResource):
-        prefix: str
-
-        def output(self, text: str) -> None:
-            out_txt.append(f"{self.prefix}{text}")
-
-    @op
-    def hello_world_op(writer: WriterResource):
-        writer.output("hello, world!")
-
-    @job
-    def hello_world_job():
-        hello_world_op()
-
-    @sensor(job=hello_world_job)
-    def hello_world_sensor():
-        ...
-
-    hello_world_schedule = ScheduleDefinition(
-        name="hello_world_schedule", cron_schedule="* * * * *", job=hello_world_job
-    )
-
-    # Bind the resource to the job at definition time and validate that it works
-    defs = Definitions(
-        jobs=[hello_world_job],
-        schedules=[hello_world_schedule],
-        sensors=[hello_world_sensor],
-        resources={
-            "writer": WriterResource(prefix="msg: "),
-        },
-    )
-
-    assert (
-        cast(JobDefinition, defs.get_sensor_def("hello_world_sensor").job)
-        .execute_in_process()
-        .success
-    )
-    assert out_txt == ["msg: hello, world!"]
-
-    out_txt.clear()
-
-    assert (
-        cast(JobDefinition, defs.get_schedule_def("hello_world_schedule").job)
-        .execute_in_process()
-        .success
-    )
-    assert out_txt == ["msg: hello, world!"]
-
-    out_txt.clear()
-
-
-def test_bind_resource_to_instigator_by_name() -> None:
-    out_txt = []
-
-    class WriterResource(ConfigurableResource):
-        prefix: str
-
-        def output(self, text: str) -> None:
-            out_txt.append(f"{self.prefix}{text}")
-
-    @op
-    def hello_world_op(writer: WriterResource):
-        writer.output("hello, world!")
-
-    @job
-    def hello_world_job():
-        hello_world_op()
-
-    @sensor(job_name="hello_world_job")
-    def hello_world_sensor():
-        ...
-
-    hello_world_schedule = ScheduleDefinition(
-        name="hello_world_schedule", cron_schedule="* * * * *", job_name="hello_world_job"
-    )
-
-    # Bind the resource to the job at definition time and validate that it works
-    defs = Definitions(
-        jobs=[hello_world_job],
-        schedules=[hello_world_schedule],
-        sensors=[hello_world_sensor],
-        resources={
-            "writer": WriterResource(prefix="msg: "),
-        },
-    )
-
-    assert (
-        defs.get_job_def(cast(str, defs.get_sensor_def("hello_world_sensor").job_name))
-        .execute_in_process()
-        .success
-    )
-    assert out_txt == ["msg: hello, world!"]
-
-    out_txt.clear()
-
-    assert (
-        defs.get_job_def(cast(str, defs.get_schedule_def("hello_world_schedule").job_name))
-        .execute_in_process()
-        .success
-    )
-    assert out_txt == ["msg: hello, world!"]
-
-    out_txt.clear()
-
-
-def test_bind_io_manager_default() -> None:
-    outputs = []
-
-    class MyIOManager(ConfigurableIOManager):
-        def load_input(self, _) -> None:
-            pass
-
-        def handle_output(self, _, obj) -> None:
-            outputs.append(obj)
-
-    @op
-    def hello_world_op() -> str:
-        return "foo"
-
-    @job
-    def hello_world_job() -> None:
-        hello_world_op()
-
-    # Bind the I/O manager to the job at definition time and validate that it works
-    defs = Definitions(
-        jobs=[hello_world_job],
-        resources={
-            "io_manager": MyIOManager(),
-        },
-    )
-
-    assert defs.get_job_def("hello_world_job").execute_in_process().success
-    assert outputs == ["foo"]
-
-
-def test_bind_io_manager_override() -> None:
-    outputs = []
-
-    class MyIOManager(ConfigurableIOManager):
-        def load_input(self, _) -> None:
-            pass
-
-        def handle_output(self, _, obj) -> None:
-            outputs.append(obj)
-
-    class MyOtherIOManager(ConfigurableIOManager):
-        def load_input(self, _) -> None:
-            pass
-
-        def handle_output(self, _, obj) -> None:
-            pass
-
-    @op
-    def hello_world_op() -> str:
-        return "foo"
-
-    @job(resource_defs={"io_manager": MyIOManager()})
-    def hello_world_job() -> None:
-        hello_world_op()
-
-    # Bind the I/O manager to the job at definition time and validate that it does
-    # not take precedence over the one defined on the job
-    defs = Definitions(
-        jobs=[hello_world_job],
-        resources={
-            "io_manager": MyOtherIOManager(),
-        },
-    )
-
-    assert defs.get_job_def("hello_world_job").execute_in_process().success
-    assert outputs == ["foo"]
 
 
 def test_aliased_field_structured_resource():
@@ -1946,284 +1176,6 @@ def test_aliased_field_structured_resource():
         {"resources": {"writer": {"config": {"prefix": "runtime: "}}}}
     ).success
     assert out_txt == ["runtime: hello, world!"]
-
-
-def test_direct_op_invocation() -> None:
-    class MyResource(ConfigurableResource):
-        a_str: str
-
-    @op
-    def my_op(context, my_resource: MyResource) -> str:
-        assert my_resource.a_str == "foo"
-        return my_resource.a_str
-
-    # Just providing context is ok, we'll use the resource from the context
-    assert my_op(build_op_context(resources={"my_resource": MyResource(a_str="foo")})) == "foo"
-
-    # Providing both context and resource is not ok, because we don't know which one to use
-    with pytest.raises(
-        DagsterInvalidInvocationError,
-        match="Cannot provide resources in both context and kwargs",
-    ):
-        assert (
-            my_op(
-                context=build_op_context(resources={"my_resource": MyResource(a_str="foo")}),
-                my_resource=MyResource(a_str="foo"),
-            )
-            == "foo"
-        )
-
-    # Providing resource only as kwarg is ok, we'll use that (we still need a context though)
-    assert my_op(context=build_op_context(), my_resource=MyResource(a_str="foo")) == "foo"
-
-    # Providing resource only as positional arg is ok, we'll use that (we still need a context though)
-    assert my_op(build_op_context(), MyResource(a_str="foo")) == "foo"
-
-    @op
-    def my_op_no_context(my_resource: MyResource) -> str:
-        assert my_resource.a_str == "foo"
-        return my_resource.a_str
-
-    # Providing context is ok, we just discard it and use the resource from the context
-    assert (
-        my_op_no_context(build_op_context(resources={"my_resource": MyResource(a_str="foo")}))
-        == "foo"
-    )
-
-    # Providing resource only as kwarg is ok, we'll use that
-    assert my_op_no_context(my_resource=MyResource(a_str="foo")) == "foo"
-
-
-def test_direct_op_invocation_multiple_resources() -> None:
-    class MyResource(ConfigurableResource):
-        a_str: str
-
-    @op
-    def my_op(context, my_resource: MyResource, my_other_resource: MyResource) -> str:
-        assert my_resource.a_str == "foo"
-        assert my_other_resource.a_str == "bar"
-        return my_resource.a_str
-
-    # Just providing context is ok, we'll use both resources from the context
-    assert (
-        my_op(
-            build_op_context(
-                resources={
-                    "my_resource": MyResource(a_str="foo"),
-                    "my_other_resource": MyResource(a_str="bar"),
-                }
-            )
-        )
-        == "foo"
-    )
-
-    # Providing resource only as kwarg is ok, we'll use that (we still need a context though)
-    assert (
-        my_op(
-            context=build_op_context(),
-            my_resource=MyResource(a_str="foo"),
-            my_other_resource=MyResource(a_str="bar"),
-        )
-        == "foo"
-    )
-
-    @op
-    def my_op_no_context(my_resource: MyResource, my_other_resource: MyResource) -> str:
-        assert my_resource.a_str == "foo"
-        assert my_other_resource.a_str == "bar"
-        return my_resource.a_str
-
-    # Providing context is ok, we just discard it and use the resource from the context
-    assert (
-        my_op_no_context(
-            build_op_context(
-                resources={
-                    "my_resource": MyResource(a_str="foo"),
-                    "my_other_resource": MyResource(a_str="bar"),
-                }
-            )
-        )
-        == "foo"
-    )
-
-    # Providing resource only as kwarg is ok, we'll use that
-    assert (
-        my_op_no_context(
-            my_resource=MyResource(a_str="foo"), my_other_resource=MyResource(a_str="bar")
-        )
-        == "foo"
-    )
-
-
-def test_direct_op_invocation_with_inputs() -> None:
-    class MyResource(ConfigurableResource):
-        z: int
-
-    @op
-    def my_wacky_addition_op(context, x: int, y: int, my_resource: MyResource) -> int:
-        return x + y + my_resource.z
-
-    # Just providing context is ok, we'll use the resource from the context
-    # We are successfully able to input x and y as args
-    assert (
-        my_wacky_addition_op(build_op_context(resources={"my_resource": MyResource(z=2)}), 4, 5)
-        == 11
-    )
-    # We can also input x and y as kwargs
-    assert (
-        my_wacky_addition_op(build_op_context(resources={"my_resource": MyResource(z=3)}), y=1, x=2)
-        == 6
-    )
-
-    # Providing resource only as kwarg is ok, we'll use that (we still need a context though)
-    # We can input x and y as args
-    assert my_wacky_addition_op(build_op_context(), 10, 20, my_resource=MyResource(z=30)) == 60
-    # We can also input x and y as kwargs in this case
-    assert my_wacky_addition_op(build_op_context(), y=1, x=2, my_resource=MyResource(z=3)) == 6
-
-    @op
-    def my_wacky_addition_op_no_context(x: int, y: int, my_resource: MyResource) -> int:
-        return x + y + my_resource.z
-
-    # Providing context is ok, we just discard it and use the resource from the context
-    # We can input x and y as args
-    assert (
-        my_wacky_addition_op_no_context(
-            build_op_context(resources={"my_resource": MyResource(z=2)}), 4, 5
-        )
-        == 11
-    )
-    # We can also input x and y as kwargs
-    assert (
-        my_wacky_addition_op_no_context(
-            build_op_context(resources={"my_resource": MyResource(z=3)}), y=1, x=2
-        )
-        == 6
-    )
-
-    # Providing resource only as kwarg is ok, we'll use that
-    # We can input x and y as args
-    assert my_wacky_addition_op_no_context(10, 20, my_resource=MyResource(z=30)) == 60
-    # We can also input x and y as kwargs in this case
-    assert my_wacky_addition_op_no_context(y=1, x=2, my_resource=MyResource(z=3)) == 6
-
-    # Direct invocation is a little weird if the resource comes before an input,
-    # but it still works as long as you use kwargs for the inputs or provide the resource explicitly
-    @op
-    def my_wacky_addition_op_resource_first(my_resource: MyResource, x: int, y: int) -> int:
-        return x + y + my_resource.z
-
-    # Here we have to use kwargs for x and y because we're not providing the resource explicitly
-    assert (
-        my_wacky_addition_op_resource_first(
-            build_op_context(resources={"my_resource": MyResource(z=2)}), x=4, y=5
-        )
-        == 11
-    )
-
-    # Here we can just use args for x and y because we're providing the resource explicitly as an arg
-    assert my_wacky_addition_op_resource_first(MyResource(z=2), 45, 53) == 100
-
-
-def test_direct_asset_invocation() -> None:
-    class MyResource(ConfigurableResource):
-        a_str: str
-
-    @asset
-    def my_asset(context, my_resource: MyResource) -> str:
-        assert my_resource.a_str == "foo"
-        return my_resource.a_str
-
-    # Just providing context is ok, we'll use the resource from the context
-    assert my_asset(build_op_context(resources={"my_resource": MyResource(a_str="foo")})) == "foo"
-
-    # Providing both context and resource is not ok, because we don't know which one to use
-    with pytest.raises(
-        DagsterInvalidInvocationError,
-        match="Cannot provide resources in both context and kwargs",
-    ):
-        assert (
-            my_asset(
-                context=build_op_context(resources={"my_resource": MyResource(a_str="foo")}),
-                my_resource=MyResource(a_str="foo"),
-            )
-            == "foo"
-        )
-
-    # Providing resource only as kwarg is ok, we'll use that (we still need a context though)
-    assert my_asset(context=build_op_context(), my_resource=MyResource(a_str="foo")) == "foo"
-
-    # Providing resource  as arg is ok, we'll use that (we still need a context though)
-    assert my_asset(build_op_context(), MyResource(a_str="foo")) == "foo"
-
-    @asset
-    def my_asset_no_context(my_resource: MyResource) -> str:
-        assert my_resource.a_str == "foo"
-        return my_resource.a_str
-
-    # Providing context is ok, we just discard it and use the resource from the context
-    assert (
-        my_asset_no_context(build_op_context(resources={"my_resource": MyResource(a_str="foo")}))
-        == "foo"
-    )
-
-    # Providing resource only as kwarg is ok, we'll use that
-    assert my_asset_no_context(my_resource=MyResource(a_str="foo")) == "foo"
-
-
-def test_direct_asset_invocation_with_inputs() -> None:
-    class MyResource(ConfigurableResource):
-        z: int
-
-    @asset
-    def my_wacky_addition_asset(context, x: int, y: int, my_resource: MyResource) -> int:
-        return x + y + my_resource.z
-
-    # Just providing context is ok, we'll use the resource from the context
-    # We are successfully able to input x and y as args
-    assert (
-        my_wacky_addition_asset(build_op_context(resources={"my_resource": MyResource(z=2)}), 4, 5)
-        == 11
-    )
-    # We can also input x and y as kwargs
-    assert (
-        my_wacky_addition_asset(
-            build_op_context(resources={"my_resource": MyResource(z=3)}), y=1, x=2
-        )
-        == 6
-    )
-
-    # Providing resource only as kwarg is ok, we'll use that (we still need a context though)
-    # We can input x and y as args
-    assert my_wacky_addition_asset(build_op_context(), 10, 20, my_resource=MyResource(z=30)) == 60
-    # We can also input x and y as kwargs in this case
-    assert my_wacky_addition_asset(build_op_context(), y=1, x=2, my_resource=MyResource(z=3)) == 6
-
-    @asset
-    def my_wacky_addition_asset_no_context(x: int, y: int, my_resource: MyResource) -> int:
-        return x + y + my_resource.z
-
-    # Providing context is ok, we just discard it and use the resource from the context
-    # We can input x and y as args
-    assert (
-        my_wacky_addition_asset_no_context(
-            build_op_context(resources={"my_resource": MyResource(z=2)}), 4, 5
-        )
-        == 11
-    )
-    # We can also input x and y as kwargs
-    assert (
-        my_wacky_addition_asset_no_context(
-            build_op_context(resources={"my_resource": MyResource(z=3)}), y=1, x=2
-        )
-        == 6
-    )
-
-    # Providing resource only as kwarg is ok, we'll use that
-    # We can input x and y as args
-    assert my_wacky_addition_asset_no_context(10, 20, my_resource=MyResource(z=30)) == 60
-    # We can also input x and y as kwargs in this case
-    assert my_wacky_addition_asset_no_context(y=1, x=2, my_resource=MyResource(z=3)) == 6
 
 
 def test_from_resource_context_and_to_config_field() -> None:
@@ -2272,131 +1224,6 @@ def test_from_resource_context_and_to_config_field_complex() -> None:
         "a": [{"b": 1.0}, {"c": 2.0}],
         "d": [{"e": 3.0}, {"f": 4.0}],
     }
-
-
-def test_direct_op_invocation_plain_arg_with_resource_definition_no_inputs_no_context() -> None:
-    class NumResource(ConfigurableResource):
-        num: int
-
-    executed = {}
-
-    @op
-    def an_op(my_resource: NumResource) -> None:
-        assert my_resource.num == 1
-        executed["yes"] = True
-
-    an_op(NumResource(num=1))
-
-    assert executed["yes"]
-
-
-def test_direct_op_invocation_kwarg_with_resource_definition_no_inputs_no_context() -> None:
-    class NumResource(ConfigurableResource):
-        num: int
-
-    executed = {}
-
-    @op
-    def an_op(my_resource: NumResource) -> None:
-        assert my_resource.num == 1
-        executed["yes"] = True
-
-    an_op(my_resource=NumResource(num=1))
-
-    assert executed["yes"]
-
-
-def test_direct_asset_invocation_plain_arg_with_resource_definition_no_inputs_no_context() -> None:
-    class NumResource(ConfigurableResource):
-        num: int
-
-    executed = {}
-
-    @asset
-    def an_asset(my_resource: NumResource) -> None:
-        assert my_resource.num == 1
-        executed["yes"] = True
-
-    an_asset(NumResource(num=1))
-
-    assert executed["yes"]
-
-
-def test_direct_asset_invocation_kwarg_with_resource_definition_no_inputs_no_context() -> None:
-    class NumResource(ConfigurableResource):
-        num: int
-
-    executed = {}
-
-    @asset
-    def an_asset(my_resource: NumResource) -> None:
-        assert my_resource.num == 1
-        executed["yes"] = True
-
-    an_asset(my_resource=NumResource(num=1))
-
-    assert executed["yes"]
-
-
-def test_direct_asset_invocation_many_resource_args() -> None:
-    class NumResource(ConfigurableResource):
-        num: int
-
-    executed = {}
-
-    @asset
-    def an_asset(my_resource: NumResource, my_other_resource: NumResource) -> None:
-        assert my_resource.num == 1
-        assert my_other_resource.num == 2
-        executed["yes"] = True
-
-    an_asset(NumResource(num=1), NumResource(num=2))
-    assert executed["yes"]
-    executed.clear()
-
-    an_asset(my_resource=NumResource(num=1), my_other_resource=NumResource(num=2))
-    assert executed["yes"]
-    executed.clear()
-
-    an_asset(my_other_resource=NumResource(num=2), my_resource=NumResource(num=1))
-    assert executed["yes"]
-    executed.clear()
-
-    an_asset(NumResource(num=1), my_other_resource=NumResource(num=2))
-    assert executed["yes"]
-
-
-def test_direct_asset_invocation_many_resource_args_context() -> None:
-    class NumResource(ConfigurableResource):
-        num: int
-
-    executed = {}
-
-    @asset
-    def an_asset(context, my_resource: NumResource, my_other_resource: NumResource) -> None:
-        assert context.resources.my_resource.num == 1
-        assert context.resources.my_other_resource.num == 2
-        assert my_resource.num == 1
-        assert my_other_resource.num == 2
-        executed["yes"] = True
-
-    an_asset(build_op_context(), NumResource(num=1), NumResource(num=2))
-    assert executed["yes"]
-    executed.clear()
-
-    an_asset(
-        build_op_context(), my_resource=NumResource(num=1), my_other_resource=NumResource(num=2)
-    )
-    assert executed["yes"]
-    executed.clear()
-
-    an_asset(
-        my_other_resource=NumResource(num=2),
-        my_resource=NumResource(num=1),
-        context=build_op_context(),
-    )
-    assert executed["yes"]
-    executed.clear()
 
 
 def test_from_resource_context_and_to_config_empty() -> None:
@@ -2566,35 +1393,3 @@ def test_context_on_resource_nested() -> None:
 
         assert defs.get_implicit_global_asset_job_def().execute_in_process().success
         assert executed["yes"]
-
-
-def test_bind_top_level_resource_sensor_multi_job() -> None:
-    executed = {}
-
-    class FooResource(ConfigurableResource):
-        my_str: str
-
-    @op
-    def hello_world_op(foo: FooResource):
-        assert foo.my_str == "foo"
-        executed["yes"] = True
-
-    @job()
-    def hello_world_job():
-        hello_world_op()
-
-    @job
-    def hello_world_job_2():
-        hello_world_op()
-
-    @sensor(jobs=[hello_world_job, hello_world_job_2])
-    def hello_world_sensor(context):
-        return RunRequest(run_key="foo")
-
-    Definitions(
-        sensors=[hello_world_sensor],
-        jobs=[hello_world_job, hello_world_job_2],
-        resources={
-            "foo": FooResource(my_str="foo"),
-        },
-    )
